Add shims to work against both Spark 3.5 and 4.0
